+++ scylla/src/connection/connection_pool.rs
@@ -1,22 +1,19 @@
 #[cfg(feature = "cloud")]
 use crate::cloud::set_ssl_config_for_scylla_cloud_host;
 
+use crate::connection::{self, Connection, ConnectionConfig, ErrorReceiver, VerifiedKeyspaceName};
 use crate::routing::{Shard, ShardCount, Sharder};
 use crate::transport::errors::{
     BrokenConnectionErrorKind, ConnectionError, ConnectionPoolError, QueryError,
 };
-use crate::transport::{
-    connection,
-    connection::{Connection, ConnectionConfig, ErrorReceiver, VerifiedKeyspaceName},
-};
 
 #[cfg(feature = "cloud")]
-use super::node::resolve_hostname;
+use crate::transport::node::resolve_hostname;
 
-use super::metadata::{PeerEndpoint, UntranslatedEndpoint};
+use crate::transport::metadata::{PeerEndpoint, UntranslatedEndpoint};
 #[cfg(feature = "cloud")]
-use super::node::ResolvedContactPoint;
-use super::NodeAddr;
+use crate::transport::node::ResolvedContactPoint;
+use crate::transport::NodeAddr;
 
 use arc_swap::ArcSwap;
 use futures::{future::RemoteHandle, stream::FuturesUnordered, Future, FutureExt, StreamExt};
@@ -1107,7 +1104,7 @@ impl PoolRefiller {
             .await
             .map_err(|_| QueryError::TimeoutError)?;
 
-            super::cluster::use_keyspace_result(use_keyspace_results.into_iter())
+            crate::transport::cluster::use_keyspace_result(use_keyspace_results.into_iter())
         };
 
         tokio::task::spawn(async move {
@@ -1219,10 +1216,9 @@ async fn open_connection_to_shard_aware_port(
 
 #[cfg(test)]
 mod tests {
-    use super::open_connection_to_shard_aware_port;
+    use super::{open_connection_to_shard_aware_port, ConnectionConfig};
     use crate::routing::{ShardCount, Sharder};
     use crate::test_utils::setup_tracing;
-    use crate::transport::connection::ConnectionConfig;
     use crate::transport::metadata::UntranslatedEndpoint;
     use crate::transport::node::ResolvedContactPoint;
     use std::net::{SocketAddr, ToSocketAddrs};
