+++ scylla/src/transport/cluster.rs
@@ -1,3 +1,4 @@
+use crate::connection::{Connection, PoolConfig, VerifiedKeyspaceName};
 /// Cluster manages up to date information and connections to database nodes
 use crate::frame::response::event::{Event, StatusChangeEvent};
 use crate::prepared_statement::TokenCalculationError;
@@ -6,8 +7,6 @@ use crate::session::TABLET_CHANNEL_SIZE;
 use crate::transport::errors::{BadQuery, NewSessionError, QueryError};
 use crate::transport::host_filter::HostFilter;
 use crate::transport::{
-    connection::{Connection, VerifiedKeyspaceName},
-    connection_pool::PoolConfig,
     metadata::{Keyspace, Metadata, MetadataReader},
     node::Node,
     partitioner::PartitionerName,
