+++ scylla/src/transport/iterator.rs
@@ -19,13 +19,15 @@ use std::result::Result;
 use thiserror::Error;
 use tokio::sync::mpsc;
 
+use super::errors::UserRequestError;
 use super::execution_profile::ExecutionProfileInner;
 use super::query_result::ColumnSpecs;
 #[allow(deprecated)]
 use crate::cql_to_rust::{FromRow, FromRowError};
+use crate::deserialize::DeserializeOwnedRow;
 use crate::session::RequestSpan;
 
-use crate::deserialize::DeserializeOwnedRow;
+use crate::connection::{Connection, NonErrorQueryResponse, QueryResponse};
 use crate::frame::response::{
     result,
     result::{ColumnSpec, Row},
@@ -34,8 +36,7 @@ use crate::history::{self, HistoryListener};
 use crate::statement::{prepared_statement::PreparedStatement, query::Query};
 use crate::statement::{Consistency, PagingState, SerialConsistency};
 use crate::transport::cluster::ClusterData;
-use crate::transport::connection::{Connection, NonErrorQueryResponse, QueryResponse};
-use crate::transport::errors::{ProtocolError, QueryError, UserRequestError};
+use crate::transport::errors::{ProtocolError, QueryError};
 use crate::transport::load_balancing::{self, RoutingInfo};
 use crate::transport::metrics::Metrics;
 use crate::transport::retry_policy::{QueryInfo, RetryDecision, RetrySession};
